Run sysupgrade on build.
Without running sysupgrade, there is the potential for odd "missed"
dependencies (not really missed as arch guidelines specify to not
include base and base-devel dependencies), that can cause commands not
to install properly.

The filesystem package can't be updated right now.  It attempts to write
to /etc/hosts and /etc/resolv.conf which are read-only in Docker right
now.  Future versions of docker may relax this or include the
"filesystem" ignore by default, so it's best to wait it out to see what
they do.