Update Dockerfile to get build working.
This is a bit of a nasty commit, but this code was fairly out of date so
meh for now.  Pulling down updtaed mirror list to make sure we don't
have old mirrors and refreshing pacman keys to make sure things are good
there.  Also doing a force update on openssl - pacman was failing with
the old openssl.  Finally, pulling pkgbuild-introspection from main repo
rather than the AUR git one.  Things should be updated enough there for
this to work, and that's part of the official AUR guide now.